NumberBox updates
