+++ FluentAvalonia/UI/Controls/NumberBox/NumberBoxParser.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Globalization;
-using System.Text;
 using System.Text.RegularExpressions;
 
 namespace FluentAvalonia.UI.Controls
 {
-    //TODO update this to use spans... & use double?
     internal static class NumberBoxParser
     {
         public const string numberBoxOperators = "+-*/^";
